+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Changes
+
+* [#591](https://github.com/rubocop/rubocop-rails/issues/591): Add `change_column` to `Rails/ReversibleMigration`. ([@mattmccormick][])
+
 ## 2.12.4 (2021-10-16)
 
 ### Bug fixes
@@ -484,3 +488,4 @@
 [@theunraveler]: https://github.com/theunraveler
 [@pirj]: https://github.com/pirj
 [@vitormd]: https://github.com/vitormd
+[@mattmccormick]: https://github.com/mattmccormick